#11 refctoring and errors fixing
