+++ Task-5/src/App.jsx
@@ -1,5 +1,5 @@
-import "./styles.css";
-import Course from './Course'
+import './styles.css';
+import Course from './components/Course';
 
 const App = () => {
   const courses = [
@@ -10,25 +10,25 @@ const App = () => {
         {
           name: 'Fundamentals of React',
           exercises: 10,
-          id: 1
+          id: 1,
         },
         {
           name: 'Using props to pass data',
           exercises: 7,
-          id: 2
+          id: 2,
         },
         {
           name: 'State of a component',
           exercises: 14,
-          id: 3
+          id: 3,
         },
         {
           name: 'Redux',
           exercises: 11,
-          id: 4
-        }
-      ]
-    }, 
+          id: 4,
+        },
+      ],
+    },
     {
       name: 'Node.js',
       id: 2,
@@ -36,22 +36,22 @@ const App = () => {
         {
           name: 'Routing',
           exercises: 3,
-          id: 1
+          id: 1,
         },
         {
           name: 'Middlewares',
           exercises: 7,
-          id: 2
-        }
-      ]
-    }
-  ]
+          id: 2,
+        },
+      ],
+    },
+  ];
 
   return (
     <div>
-      <Course courses={courses}/>
+      <Course courses={courses} />
     </div>
-  )
-}
+  );
+};
 
-export default App
+export default App;
