+++ Task-5/src/components/Total.jsx
@@ -1,0 +1,12 @@
+const Total = (props) => {
+  const { parts } = props;
+  console.log('parts', parts);
+  let totalExercises = parts?.reduce((sum, item) => sum + item.exercises, 0);
+  return (
+    <div>
+      <b>total of {totalExercises} exercises</b>
+    </div>
+  );
+};
+
+export default Total;
